requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.10.0
kfp>=2.0.0
pandas>=1.5.3
numpy>=1.23.5
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.5
imbalanced-learn>=0.10.0
joblib>=1.2.0
pandas-gbq>=0.19.0
db-dtypes>=1.1.0
pyarrow>=10.0.0
flask>=2.2.0
gunicorn>=20.1.0